Test for checking if gun hits to kill calculation function call chains works passed
Also fixed imports of the other tests.
Fixed the bug that stopped the assignment of caclulator within GunDamageInfo class.

+++ PF_DamageRangesCalculator/PF_Packages/dataTypes.py
@@ -2,16 +2,18 @@
 import enum
 from typing import TypeAlias, Tuple
 
+
 # Special enum for a single hit
 class Hit(enum.Enum):
     HEAD = "head"
     TORSO = "torso"
     Base = "base"
 
+
 # General type for all number of hits
 Hits: TypeAlias = Tuple[int, int, int]
 # Tuple formatted as such: (# of headshots, # of torso shots, # of limb/base shots)
 # Grenade damage counts as base damage in Hit, and for Hits would be (0, 0, 1)
 
-HitsToKill: TypeAlias = dict[str, int]
+HitsToKill: TypeAlias = dict[str, float]
 # How hits to kill are stored, with the description of number of shots, and range it can kill up to
